pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "spawn-cli"
version = "0.1.0"
description = "A minimal installable Python CLI package."
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "ruamel.yaml>=0.18",
    "filelock>=3.12",
    "httpx>=0.27",
    'tomli>=2.0; python_version < "3.11"',
    "tomli-w>=1.0",
]

[project.scripts]
spawn = "spawn_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"spawn_cli.resources" = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
pythonpath = ["src"]
